requirements.txt
streamlit>=1.20.0,<2.0
openai>=0.28.0,<1.1
python-docx>=0.8.11
PyPDF2>=3.0.0
